GOOD CATCH TLA: actors need to learn about dropped msgs
Prior to this commit, ingress actors
tracked drop counts. But we can
have the bug:

a sends m1(a) to b
b receives, b sends m2(a) to c
c receives, c sends m3(a) to b
m3(a) is dropped
a becomes idle and takes a snapshot

Looking at a's snapshot and
the ingress actor, it looks like
a's message to b was dropped.
But actually, b got the message.

Now, actor states are updated
with dropped message info.